LCH-6356: Created conditional common trait.
